requirements.txt
streamlit
numpy
scikit-image
matplotlib
Image